# Create explicit Token constructors, Token and Marker as struct