--- a/src/Esprima/Marker.cs
+++ b/src/Esprima/Marker.cs
@@ -1,19 +1,3 @@
 ﻿namespace Esprima;
 
-internal sealed class Marker
-{
-    public int Index;
-    public int Line;
-    public int Column;
-
-    public Marker()
-    {
-    }
-
-    public Marker(int index, int line, int column)
-    {
-        Index = index;
-        Line = line;
-        Column = column;
-    }
-}
+internal readonly record struct Marker(int Index, int Line, int Column);