… src/Esprima/JsxToken.cs
… src/Esprima/Scanner.cs
… src/Esprima/Token.cs
… test/Esprima.Tests/Esprima.Tests.csproj
… test/Esprima.Tests/Fixtures.cs