* factory constructors initialize use case specific fields
* trim Token fields to minimize count and size
* change Token to be a read-only struct

Jint tests are unaffected by this.