@jogibear9988 2022-08-10
Is a struct really preferable?

When I read here:
https://docs.microsoft.com/en-us/dotnet/standard/design-guidelines/choosing-between-class-and-struct

- It has an instance size under 16 bytes.

If I look at your code:

        public readonly struct Token
        {
            private Token(
                TokenType type,
                object? value,
                int start,
                int end,
                int lineNumber, 
                int lineStart,
                bool octal = false,
                char notEscapeSequenceHead = (char) 0,
                bool head = false,
                bool tail = false, 
                object? customValue = null)

these are 11 Properties, 4x an int32 wich are already 16 bytes, two pointers = 8 or 16bytes, 3 bools, 1 enum, 1char = 5bytes, 
so we have already 29bytes (or 37 if a pointer is 8 byte)