@lahma 2022-08-14
Here are the first results for struct-based solution.

``` ini
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22621
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=6.0.400
  [Host]     : .NET 6.0.8 (6.0.822.36306), X64 RyuJIT
  DefaultJob : .NET 6.0.8 (6.0.822.36306), X64 RyuJIT
```

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|30.478 ms|0.3480 ms|1687.5000|968.7500|343.7500|24 MB|
| **New** |	|	| **26.157 ms (-14%)** | **0.2075 ms** | **1000.0000 (-41%)** | **562.5000 (-42%)** | **187.5000 (-45%)** | **14 MB (-42%)** |
| Old |ParseProgram|backbone-1.1.0|3.167 ms|0.0034 ms|257.8125|136.7188|54.6875|4 MB|
| **New** |	|	| **3.143 ms (-1%)** | **0.0310 ms** | **160.1563 (-38%)** | **93.7500 (-31%)** | **31.2500 (-43%)** | **2 MB (-50%)** |
| Old |ParseProgram|jquery-1.9.1|23.224 ms|0.2257 ms|1406.2500|843.7500|312.5000|20 MB|
| **New** |	|	| **20.214 ms (-13%)** | **0.2625 ms** | **843.7500 (-40%)** | **500.0000 (-41%)** | **187.5000 (-40%)** | **12 MB (-40%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|46.191 ms|0.8820 ms|2272.7273|1363.6364|454.5455|33 MB|
| **New** |	|	| **33.333 ms (-28%)** | **0.4756 ms** | **1266.6667 (-44%)** | **733.3333 (-46%)** | **333.3333 (-27%)** | **19 MB (-42%)** |
| Old |ParseProgram|mootools-1.4.5|14.784 ms|0.1339 ms|1109.3750|578.1250|234.3750|17 MB|
| **New** |	|	| **15.343 ms (+4%)** | **0.0195 ms** | **703.1250 (-37%)** | **421.8750 (-27%)** | **203.1250 (-13%)** | **10 MB (-41%)** |
| Old |ParseProgram|underscore-1.5.2|1.419 ms|0.0037 ms|199.2188|99.6094|-|3 MB|
| **New** |	|	| **1.541 ms (+9%)** | **0.0055 ms** | **117.1875 (-41%)** | **56.6406 (-43%)** | **-** | **2 MB (-33%)** |
| Old |ParseProgram|yui-3.12.0|17.831 ms|0.2276 ms|1093.7500|625.0000|218.7500|16 MB|
| **New** |	|	| **13.111 ms (-26%)** | **0.0965 ms** | **671.8750 (-39%)** | **437.5000 (-30%)** | **203.1250 (-7%)** | **9 MB (-44%)** |

@jogibear9988 I believe the documented rules are for generic types passed around, like `DateTime` etc. In this case Token is very short lived and a lot of them are being created and then just tossed away. The numbers pretty much speak for themselves I guess.