@lahma 2022-08-14
@adams85 please take a look, I know that the `TokenType` is now mixing again JSX concepts, but this solution keeps the struct small and manageable without tricks.