@lahma 2022-08-14
I also converted Marker to struct and some `in` usage to avoid copies, which gave further improvements. The benchmark has collect tokens = true, which creates unnecessary red herrings about memory usage (now that `ParsedToken` is separate, it was wasy to spot), but should be switched to off as a separate task.

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|30.478 ms|0.3480 ms|1687.5000|968.7500|343.7500|24 MB|
| **New** |	|	| **24.277 ms (-20%)** | **0.1647 ms** | **937.5000 (-44%)** | **562.5000 (-42%)** | **187.5000 (-45%)** | **13 MB (-46%)** |
| Old |ParseProgram|backbone-1.1.0|3.167 ms|0.0034 ms|257.8125|136.7188|54.6875|4 MB|
| **New** |	|	| **2.939 ms (-7%)** | **0.0333 ms** | **144.5313 (-44%)** | **93.7500 (-31%)** | **31.2500 (-43%)** | **2 MB (-50%)** |
| Old |ParseProgram|jquery-1.9.1|23.224 ms|0.2257 ms|1406.2500|843.7500|312.5000|20 MB|
| **New** |	|	| **19.963 ms (-14%)** | **0.3915 ms** | **750.0000 (-47%)** | **437.5000 (-48%)** | **156.2500 (-50%)** | **11 MB (-45%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|46.191 ms|0.8820 ms|2272.7273|1363.6364|454.5455|33 MB|
| **New** |	|	| **25.223 ms (-45%)** | **0.3831 ms** | **1312.5000 (-42%)** | **843.7500 (-38%)** | **406.2500 (-11%)** | **17 MB (-48%)** |
| Old |ParseProgram|mootools-1.4.5|14.784 ms|0.1339 ms|1109.3750|578.1250|234.3750|17 MB|
| **New** |	|	| **14.424 ms (-2%)** | **0.1609 ms** | **640.6250 (-42%)** | **421.8750 (-27%)** | **203.1250 (-13%)** | **9 MB (-47%)** |
| Old |ParseProgram|underscore-1.5.2|1.419 ms|0.0037 ms|199.2188|99.6094|-|3 MB|
| **New** |	|	| **1.454 ms (+2%)** | **0.0059 ms** | **105.4688 (-47%)** | **35.1563 (-65%)** | **-** | **2 MB (-33%)** |
| Old |ParseProgram|yui-3.12.0|17.831 ms|0.2276 ms|1093.7500|625.0000|218.7500|16 MB|
| **New** |	|	| **13.457 ms (-25%)** | **0.2148 ms** | **625.0000 (-43%)** | **437.5000 (-30%)** | **203.1250 (-7%)** | **8 MB (-50%)** |